*: spring cleaning!
added a "community" section for the (new!) IRC channel and the
de facto officially-unofficial Discord channel. the primary
motivation is to not clutter up #openmpt with stuff about
schism. people have actually tried joining #schism so I think
it's a good pick for a channel name.

I spread out all of the builds again, since it was admittedly
kind of stupid to separate them all out into "sections" when
that's only even relevant for Windows.